pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "intra-trading-system"
version = "0.2.0"
description = "Trading strategy research workspace — QuantConnect + Azure hybrid"
requires-python = ">=3.11"
dependencies = [
    # Data & Analysis
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "yfinance>=0.2.30",

    # Models
    "lightgbm>=4.1.0",
    "xgboost>=2.0.0",
    "scikit-learn>=1.3.0",

    # Config & Utils
    "pyyaml>=6.0.1",
    "pytz>=2023.3",

    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
# Azure integration (for model training at scale + custom data)
azure = [
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]
# Jupyter notebooks
notebooks = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
]
# NLP & alternative data
nlp = [
    "transformers>=4.36.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
]
# Backtesting (standalone, outside QC)
backtest = [
    "vectorbt>=0.26.0",
]
# Development
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
# All optional dependencies
all = [
    "intra-trading-system[azure,notebooks,nlp,backtest,dev]",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
